requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.12.14
cloudinary==1.42.1
dj-database-url==2.3.0
Django==4.2.17
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
pillow==10.4.0
psycopg2-binary==2.9.10
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.12.2
urllib3==2.2.3
